Moved shop data structures to `shop.c/.h`
