+++ include/game.h
@@ -1,7 +1,6 @@
 #ifndef GAME_H
 #define GAME_H
 
-#include "bitset.h"
 #include "game/common_ui.h"
 #include "game_variables.h"
 #include "graphic_utils.h"
@@ -89,9 +88,6 @@ void remove_owned_joker(int owned_joker_idx);
 List* get_jokers_list(void);
 List* get_expired_jokers_list(void);
 List* get_discarded_jokers_list(void);
-List* get_shop_jokers_list(void);
-Bitset* get_avail_jokers_bitset(void);
-void set_shop_joker_avail(int joker_id, bool avail);
 
 int get_deck_top(void);
 int get_num_discards_remaining(void);
