+++ source/game/shop.c
@@ -63,6 +63,9 @@ static const Rect     SHOP_REROLL_RECT      = { 88,  96, UNDEFINED, UNDEFINED };
 static const BG_POINT SHOP_JOKER_SPRITES_INIT_POS = {120, 160};
 // clang-format on
 
+static List _shop_jokers_list = LIST_DEFAULT;
+BITSET_DEFINE(_avail_jokers_bitset, MAX_DEFINABLE_JOKERS)
+
 enum GameShopStates
 {
     GAME_SHOP_INTRO,
@@ -136,6 +139,32 @@ static int timer;
 
 static int reroll_cost = REROLL_BASE_COST;
 
+static inline void reset_shop_jokers(void)
+{
+    int num_jokers = get_joker_registry_size();
+
+    bitset_clear(&_avail_jokers_bitset);
+    for (int i = 0; i < num_jokers; i++)
+    {
+        bitset_set_idx(&_avail_jokers_bitset, i, true);
+    }
+}
+
+void game_shop_init_run(void)
+{
+    list_clear(&_shop_jokers_list);
+    _shop_jokers_list = list_init();
+    reset_shop_jokers();
+}
+
+/**
+ * @brief Set whether a Joker can appear in the shop.
+ */
+void game_shop_set_joker_avail(int joker_id, bool avail)
+{
+    bitset_set_idx(&_avail_jokers_bitset, joker_id, avail);
+}
+
 void game_shop_change_background(void)
 {
     toggle_windows(false, true);
@@ -181,7 +210,7 @@ void game_shop_on_init(void)
  */
 static inline int get_num_shop_jokers_avail(void)
 {
-    return bitset_num_set_bits(get_avail_jokers_bitset());
+    return bitset_num_set_bits(&_avail_jokers_bitset);
 }
 
 /**
@@ -203,7 +232,7 @@ static inline int game_shop_get_rand_available_joker_id(void)
     int fallback_random_joker_id = UNDEFINED;
     int match_count = 0;
 
-    BitsetItr itr = bitset_itr_create(get_avail_jokers_bitset());
+    BitsetItr itr = bitset_itr_create(&_avail_jokers_bitset);
 
     int i = 0;
     int joker_id = UNDEFINED;
@@ -229,7 +258,14 @@ static inline int game_shop_get_rand_available_joker_id(void)
  */
 static inline bool no_avail_jokers(void)
 {
-    return bitset_is_empty(get_avail_jokers_bitset());
+    return bitset_is_empty(&_avail_jokers_bitset);
+}
+
+
+GBAL_UNUSED
+static inline bool is_shop_joker_avail(int joker_id)
+{
+    return bitset_get_idx(&_avail_jokers_bitset, joker_id);
 }
 
 /**
@@ -244,7 +280,7 @@ static void game_shop_create_items(void)
     if (no_avail_jokers())
         return;
 
-    List* shop_jokers_list = get_shop_jokers_list();
+    List* shop_jokers_list = &_shop_jokers_list;
 
     list_clear(shop_jokers_list);
     *shop_jokers_list = list_init();
@@ -274,7 +310,7 @@ static void game_shop_create_items(void)
         if (joker_id == UNDEFINED)
             break;
 
-        set_shop_joker_avail(joker_id, false);
+        game_shop_set_joker_avail(joker_id, false);
 
         JokerObject* joker_object = joker_object_new(joker_new(joker_id));
 
@@ -349,7 +385,7 @@ static void game_shop_intro()
 static int shop_top_row_get_size(void)
 {
     // + 1 to account for next round button
-    return list_get_len(get_shop_jokers_list()) + 1;
+    return list_get_len(&_shop_jokers_list) + 1;
 }
 
 /**
@@ -366,7 +402,7 @@ static inline void add_to_held_jokers(JokerObject* joker_object)
  */
 static inline void game_shop_buy_joker(int shop_joker_idx)
 {
-    List* shop_jokers_list = get_shop_jokers_list();
+    List* shop_jokers_list = &_shop_jokers_list;
     JokerObject* joker_object = (JokerObject*)list_get_at_idx(shop_jokers_list, shop_joker_idx);
 
     g_game_vars.money -= joker_object->joker->value;
@@ -393,7 +429,7 @@ static void shop_top_row_on_key_transit(SelectionGrid* selection_grid, Selection
     {
         int shop_joker_idx = selection->x - 1; // - 1 to account for next round button
         JokerObject* joker_object =
-            (JokerObject*)list_get_at_idx(get_shop_jokers_list(), shop_joker_idx);
+            (JokerObject*)list_get_at_idx(&_shop_jokers_list, shop_joker_idx);
         if (joker_object == NULL || list_get_len(get_jokers_list()) >= MAX_JOKERS_HELD_SIZE ||
             g_game_vars.money < joker_object->joker->value)
         {
@@ -415,7 +451,7 @@ static bool shop_top_row_on_selection_changed(
     const Selection* new_selection
 )
 {
-    List* shop_jokers_list = get_shop_jokers_list();
+    List* shop_jokers_list = &_shop_jokers_list;
     // Guard if we move down while on jokers
     if (new_selection->y > row_idx && prev_selection->x > 0)
         return false;
@@ -482,7 +518,7 @@ static bool shop_reroll_row_on_selection_changed(
         if (new_selection->x != NEXT_ROUND_BTN_SEL_X)
         {
             int idx = new_selection->x - 1;
-            JokerObject* joker_object = (JokerObject*)list_get_at_idx(get_shop_jokers_list(), idx);
+            JokerObject* joker_object = (JokerObject*)list_get_at_idx(&_shop_jokers_list, idx);
             sprite_object_set_focus(joker_object->sprite_object, true);
         }
     }
@@ -503,7 +539,7 @@ static inline void game_shop_reroll(int* reroll_cost)
     g_game_vars.money -= *reroll_cost;
     display_money(); // Update the money display
 
-    List* shop_jokers_list = get_shop_jokers_list();
+    List* shop_jokers_list = &_shop_jokers_list;
     ListItr itr = list_itr_create(shop_jokers_list);
     JokerObject* joker_object;
 
@@ -511,7 +547,7 @@ static inline void game_shop_reroll(int* reroll_cost)
     {
         if (joker_object != NULL)
         {
-            set_shop_joker_avail(joker_object->joker->id, true);
+            game_shop_set_joker_avail(joker_object->joker->id, true);
             joker_object_destroy(&joker_object); // Destroy the joker object if it exists
         }
     }
@@ -604,7 +640,7 @@ static void game_shop_outro()
     {
         tte_erase_rect_wrapper(SHOP_PRICES_TEXT_RECT); // Erase the shop prices text
 
-        ListItr itr = list_itr_create(get_shop_jokers_list());
+        ListItr itr = list_itr_create(&_shop_jokers_list);
         JokerObject* joker_object;
         while ((joker_object = list_itr_next(&itr)))
         {
@@ -663,7 +699,7 @@ void game_shop_on_update(void)
 {
     timer++;
 
-    List* shop_jokers_list = get_shop_jokers_list();
+    List* shop_jokers_list = &_shop_jokers_list;
 
     if (!list_is_empty(shop_jokers_list))
     {
@@ -686,7 +722,7 @@ void game_shop_on_update(void)
 
 void game_shop_on_exit(void)
 {
-    List* shop_jokers_list = get_shop_jokers_list();
+    List* shop_jokers_list = &_shop_jokers_list;
     ListItr itr = list_itr_create(shop_jokers_list);
     JokerObject* joker_object;
 
@@ -695,7 +731,7 @@ void game_shop_on_exit(void)
         if (joker_object != NULL)
         {
             // Make the joker available back to shop
-            set_shop_joker_avail(joker_object->joker->id, true);
+            game_shop_set_joker_avail(joker_object->joker->id, true);
         }
         joker_object_destroy(&joker_object); // Destroy the joker objects
     }
